Move ModelsConfig form into seperate component following form convention.
